requirements.txt
pandas==1.5.3
plotly==5.13.0
PyGithub==1.57
